Fix scope and performance issue with roundResult variable and winCount()
Using the playRound() function to declare roundResult was nonsense
because that way it kept being called in showRoundResult, which caused
several issues with the following functions like the game being played
twice instead of once,results not in scope of winCount etc.

Played around a lot with different options and understanding errors and
what calls what and is in scope of which function.

Wrote a short test game of two rounds and it works and counts the wins
correctly.