Play around with layout and switch out javascript.js
Since I'm adding a lot of things that aren't in the original
assignment, I'm, writing a new script, but pilfering the old one for the
things I can reuse, so I'm keeping it for now.

Messed around with the layout a bit, getting a feel for CSS and Flexbox
and such again.